refactor: function button
